+++ app/src/main/java/eu/kanade/tachiyomi/data/track/bangumi/BangumiApi.kt
@@ -15,6 +15,7 @@ import kotlinx.serialization.json.int
 import kotlinx.serialization.json.jsonArray
 import kotlinx.serialization.json.jsonObject
 import kotlinx.serialization.json.jsonPrimitive
+import kotlinx.serialization.json.long
 import okhttp3.CacheControl
 import okhttp3.FormBody
 import okhttp3.OkHttpClient
@@ -109,7 +110,7 @@ class BangumiApi(private val client: OkHttpClient, interceptor: BangumiIntercept
 
     private fun jsonToSearch(obj: JsonObject): TrackSearch {
         return TrackSearch.create(TrackManager.BANGUMI).apply {
-            media_id = obj["id"]!!.jsonPrimitive.int
+            media_id = obj["id"]!!.jsonPrimitive.long
             title = obj["name_cn"]!!.jsonPrimitive.content
             cover_url = obj["images"]!!.jsonObject["common"]!!.jsonPrimitive.content
             summary = obj["name"]!!.jsonPrimitive.content
@@ -120,7 +121,7 @@ class BangumiApi(private val client: OkHttpClient, interceptor: BangumiIntercept
     private fun jsonToTrack(mangas: JsonObject): Track {
         return Track.create(TrackManager.BANGUMI).apply {
             title = mangas["name"]!!.jsonPrimitive.content
-            media_id = mangas["id"]!!.jsonPrimitive.int
+            media_id = mangas["id"]!!.jsonPrimitive.long
             score =
                 mangas["rating"]?.let { rating ->
                     if (rating !is JsonNull && rating is JsonObject) {
