+++ app/src/main/java/eu/kanade/tachiyomi/data/track/kitsu/KitsuApi.kt
@@ -7,10 +7,10 @@ import eu.kanade.tachiyomi.network.POST
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.buildJsonObject
-import kotlinx.serialization.json.int
 import kotlinx.serialization.json.jsonArray
 import kotlinx.serialization.json.jsonObject
 import kotlinx.serialization.json.jsonPrimitive
+import kotlinx.serialization.json.long
 import kotlinx.serialization.json.put
 import kotlinx.serialization.json.putJsonObject
 import okhttp3.FormBody
@@ -92,7 +92,7 @@ class KitsuApi(private val client: OkHttpClient, interceptor: KitsuInterceptor) 
 
             rest.addLibManga(buildJsonObject { put("data", data) })
                 .map { json ->
-                    track.media_id = json["data"]!!.jsonObject["id"]!!.jsonPrimitive.int
+                    track.media_id = json["data"]!!.jsonObject["id"]!!.jsonPrimitive.long
                     track
                 }
         }
@@ -198,20 +198,20 @@ class KitsuApi(private val client: OkHttpClient, interceptor: KitsuInterceptor) 
         @Headers("Content-Type: application/vnd.api+json")
         @PATCH("library-entries/{id}")
         fun updateLibManga(
-            @Path("id") remoteId: Int,
+            @Path("id") remoteId: Long,
             @Body data: JsonObject
         ): Observable<JsonObject>
 
         @GET("library-entries")
         fun findLibManga(
-            @Query("filter[manga_id]", encoded = true) remoteId: Int,
+            @Query("filter[manga_id]", encoded = true) remoteId: Long,
             @Query("filter[user_id]", encoded = true) userId: String,
             @Query("include") includes: String = "manga"
         ): Observable<JsonObject>
 
         @GET("library-entries")
         fun getLibManga(
-            @Query("filter[id]", encoded = true) remoteId: Int,
+            @Query("filter[id]", encoded = true) remoteId: Long,
             @Query("include") includes: String = "manga"
         ): Observable<JsonObject>
 
@@ -258,7 +258,7 @@ class KitsuApi(private val client: OkHttpClient, interceptor: KitsuInterceptor) 
         private const val algoliaAppId = "AWQO5J657S"
         private const val algoliaFilter = "&facetFilters=%5B%22kind%3Amanga%22%5D&attributesToRetrieve=%5B%22synopsis%22%2C%22canonicalTitle%22%2C%22chapterCount%22%2C%22posterImage%22%2C%22startDate%22%2C%22subtype%22%2C%22endDate%22%2C%20%22id%22%5D"
 
-        fun mangaUrl(remoteId: Int): String {
+        fun mangaUrl(remoteId: Long): String {
             return baseMangaUrl + remoteId
         }
